# core/src/main/java/com/klaimz/model/api/Filter.java
package com.klaimz.model.api;

import io.micronaut.core.annotation.Introspected;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Introspected
public class Filter {

    @NotBlank
    private String field;
    @NotBlank
    private String value;
}